+++ bin/commands/index.mjs
@@ -1,0 +1,4 @@
+import generate from './generate.mjs';
+import lint from './lint.mjs';
+
+export default [generate, lint];
